added the doc

+++ run.py
@@ -14,10 +14,30 @@ import shutil
 import docker
 
 
-##
-#
-#
 def main(inputfile, input_format, output, rbs, max_prot_per_react, tirs, topX, pathway_id):
+    """Call the docker to convert a collection of rpSBML into a collection of SBOL
+
+    :param inputfile: The path to the TAR collection of rpSBML
+    :param intput_format: The input format of the file. Valid options: tar, sbml
+    :param output: The path to the output TAR collection of SBOL
+    :param rbs: Calculate or not the RBS strength
+    :param max_prot_per_react: The maximum number of proteins per reaction (Default: 3)
+    :param tirs: The RBS strength values
+    :param topX: The top number of UNIPROT id's per reaction to use
+    :param pathway_id: The Groups id of the heterologous pathway
+
+    :type inputfile: str
+    :type input_format: str
+    :type output: str
+    :type rbs: bool
+    :type max_prot_per_react: int
+    :type tirs: list
+    :type topX: int
+    :type pathway_id: str
+
+    :rtype: None
+    :return: None
+    """
     docker_client = docker.from_env()
     image_str = 'brsynth/rpsbmltosbol-standalone'
     try:
@@ -31,42 +51,47 @@ def main(inputfile, input_format, output, rbs, max_prot_per_react, tirs, topX, p
             logging.error('Cannot pull image: '+str(image_str))
             exit(1)
     with tempfile.TemporaryDirectory() as tmpOutputFolder:
-        shutil.copy(inputfile, tmpOutputFolder+'/input.dat')
-        command = ['/home/tool_rpSBMLtoSBOL.py',
-                   '-input',
-                   '/home/tmp_output/input.dat',
-                   '-output',
-                   '/home/tmp_output/output.dat',
-                   '-input_format',
-                   str(input_format),
-                   '-pathway_id',
-                   str(pathway_id),
-                   '-rbs',
-                   str(rbs),
-                   '-topX',
-                   str(topX),
-                   '-tirs',
-                   str(tirs),
-                   '-max_prot_per_react',
-                   str(max_prot_per_react)]
-        container = docker_client.containers.run(image_str, 
-                                                 command, 
-                                                 detach=True, 
-                                                 stderr=True, 
-                                                 volumes={tmpOutputFolder+'/': {'bind': '/home/tmp_output', 'mode': 'rw'}})
-        container.wait()
-        err = container.logs(stdout=False, stderr=True)
-        err_str = err.decode('utf-8')
-        print(err_str)
-        if not 'ERROR' in err_str:
-            shutil.copy(tmpOutputFolder+'/output.dat', output)
-        container.remove()
- 
+        if os.path.exists(inputfile):
+            shutil.copy(inputfile, tmpOutputFolder+'/input.dat')
+            command = ['/home/tool_rpSBMLtoSBOL.py',
+                       '-input',
+                       '/home/tmp_output/input.dat',
+                       '-output',
+                       '/home/tmp_output/output.dat',
+                       '-input_format',
+                       str(input_format),
+                       '-pathway_id',
+                       str(pathway_id),
+                       '-rbs',
+                       str(rbs),
+                       '-topX',
+                       str(topX),
+                       '-tirs',
+                       str(tirs),
+                       '-max_prot_per_react',
+                       str(max_prot_per_react)]
+            container = docker_client.containers.run(image_str, 
+                                                     command, 
+                                                     detach=True, 
+                                                     stderr=True, 
+                                                     volumes={tmpOutputFolder+'/': {'bind': '/home/tmp_output', 'mode': 'rw'}})
+            container.wait()
+            err = container.logs(stdout=False, stderr=True)
+            err_str = err.decode('utf-8')
+            if 'ERROR' in err_str:
+                print(err_str)
+            elif 'WARNING' in err_str:
+                print(err_str)
+            if not os.path.exists(tmpOutputFolder+'/output.dat'):
+                print('ERROR: Cannot find the output file: '+str(tmpOutputFolder+'/output.dat'))
+            else:
+                shutil.copy(tmpOutputFolder+'/output.dat', output)
+            container.remove()
+        else:
+            logging.error('Cannot find one or more of the input file: '+str(inputfile))
+            exit(1)
 
 
-##
-#
-#
 if __name__ == "__main__":
     parser = argparse.ArgumentParser('Given an rpSBML or a collection of rpSBML in a tar, convert to the SBOL format based on the selenzyme UNIPROT IDs')
     parser.add_argument('-input', type=str, help='input file path', required=True)
